+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
 
 ## Improvements
 - Player One Cameras that support different sensor modes can now switch between these via ReadoutModes when using the native driver
+- Player One Cameras that have a temperature probe but not cooler will now show the temperature
 - Solve & Sync instruction will now also sync the sky angle with a rotator when connected
 
 # Version 2.0 Hotfix 2